chore: migrate to almost standard unified project
