--- .changeset/config.json
@@ -1,15 +1,0 @@
-{
-  "$schema": "https://unpkg.com/@changesets/config/schema.json",
-  "changelog": [
-    "@changesets/changelog-github",
-    {
-      "repo": "remarkjs/remark-preset-prettier"
-    }
-  ],
-  "commit": false,
-  "linked": [],
-  "access": "public",
-  "baseBranch": "master",
-  "updateInternalDependencies": "patch",
-  "ignore": []
-}
