--- .changeset/great-cherries-own.md
@@ -1,5 +1,0 @@
----
-'remark-preset-prettier': major
----
-
-feat!: use top level await and pure ESM
