--- .codesandbox/ci.json
@@ -1,4 +1,0 @@
-{
-  "node": "16",
-  "sandboxes": []
-}
