+++ .eslintrc
@@ -1,12 +1,4 @@
 {
   "root": true,
-  "extends": "@1stg",
-  "overrides": [
-    {
-      "files": ".github/*.yml",
-      "rules": {
-        "unicorn/filename-case": "off"
-      }
-    }
-  ]
+  "extends": "@1stg"
 }
