--- .github/FUNDING.yml
@@ -1,12 +1,0 @@
-github:
-  - JounQin
-  - 1stG
-  - rxts
-  - unts
-patreon: 1stG
-open_collective: unified
-custom:
-  - https://opencollective.com/1stG
-  - https://opencollective.com/rxts
-  - https://opencollective.com/unts
-  - https://afdian.net/@JounQin
