+++ .github/workflows/main.yml
@@ -1,0 +1,21 @@
+name: main
+on:
+  - pull_request
+  - push
+jobs:
+  main:
+    name: Build and Lint with Node.js ${{ matrix.node }}
+    strategy:
+      matrix:
+        node:
+          - 14
+          - 16
+          - 18
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - uses: actions/setup-node@v3
+        with:
+          node-version: ${{ matrix.node }}
+      - run: npm install
+      - run: npm run lint
