--- .github/workflows/pkg-size.yml
@@ -1,26 +1,0 @@
-name: Package Size Report
-
-on:
-  pull_request_target:
-    branches:
-      - master
-
-jobs:
-  pkg-size-report:
-    name: Package Size Report
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout
-        uses: actions/checkout@v3
-
-      - name: Setup Node.js
-        uses: actions/setup-node@v3
-        with:
-          node-version: 16
-          cache: yarn
-
-      - name: Package Size Report
-        uses: pkg-size/action@v1
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
