--- CONTRIBUTING.md
@@ -1,20 +1,0 @@
-# Contributing
-
-Hi! 👋 We’re excited that you’re interested in contributing!
-
-## Core Guidelines
-
-[Read the guidelines on the remark repository][contributing]
-
-## Packages Releasing Difference
-
-This project is managed by [changesets][], so packages releasing is controlled by [changesets][].
-
-1. There will be a release PR generated by [changesets][]'s [action](https://github.com/changesets/action) automatically, when the release is ready, you should merge that PR firstly.
-2. Then you'll need to pull the latest codes from remote.
-3. Make sure you have both GitHub repository and npm write permissions at the same time.
-4. You need a GitHub token with a `public_repo` scope as `GITHUB_TOKEN` in the environment to publish
-5. Run `yarn release` simply, or `GITHUB_TOKEN=xxx yarn release` to export `GITHUB_TOKEN` at one time.
-
-[changesets]: https://github.com/changesets/changesets
-[contributing]: https://github.com/remarkjs/.github/blob/main/contributing.md
